allow ^, &, |, and ~ on bools and vectors of bools. (also ^= &= and |=)
Why ^, &, |, and ~ instead of !=, and, or, !?

Consider:

a() and b()

If a() returns @Vector(2, bool)([_]bool{false, false}), is b() run?
How about if a() returns @Vector(2, bool)([_]bool{false, true})?

Making this defined would be slow, confusing, involve hidden control flow,
and require putting the any() all() none() branching into the language.

Even if a() and b() return "bool" these are different:

a() and b()
a() & b()

-----

I would like to throw a good error when a vector of bools is passed, but
the current architecture prevents that.